requirements.txt
torch>=2.0.0
stable-baselines3>=2.0.0
gymnasium>=0.29.0
opencv-python>=4.8.0
Pillow>=10.0.0
mss>=9.0.1
pydirectinput>=1.0.4
pynput>=1.7.6
numpy>=1.24.0
tensorboard>=2.14.0
tqdm>=4.65.0